eventRestController + map Event +adminrepotests en com + ajout d'une view detail

+++ rico-party-boot/src/main/java/rico/repo/IEventRepo.java
@@ -9,7 +9,7 @@ import rico.model.Event;
 
 
 
-public interface IEventRepo extends JpaRepository<Event, Long>{
+public interface IEventRepo extends JpaRepository<Event, Integer>{
 
 
 }
